Cechenov_ml/requirements.txt
langchain>=1.2.0
langchain-community>=0.4.0
yandexcloud>=0.379.0
faiss-cpu>=1.13.2
python-dotenv>=1.2.0
sentence-transformers>=5.2.0
SQLAlchemy>=2.0.45
